requirements.txt
pytube==10.7.2
python-dotenv==0.15.0